pyproject.toml
[project]
name = "5329a2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "open-clip-torch>=2.32.0",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "polars>=1.29.0",
    "scikit-learn>=1.6.1",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "transformers>=4.51.3",
]
